+++ internal/repository/chat_repository.go
@@ -24,7 +24,7 @@ func (r *ChatRepo) Create(chat *domain.Chat) error {
 }
 
 // FindByID finds a chat by ID
-func (r *ChatRepo) FindByID(id uint) (*domain.Chat, error) {
+func (r *ChatRepo) FindByID(id uint64) (*domain.Chat, error) {
 	var chat domain.Chat
 	err := r.db.First(&chat, id).Error
 	if err != nil {
@@ -37,14 +37,14 @@ func (r *ChatRepo) FindByID(id uint) (*domain.Chat, error) {
 }
 
 // FindByOrganizationID finds chats by organization ID with pagination
-func (r *ChatRepo) FindByOrganizationID(orgID uint, limit, offset int) ([]domain.Chat, error) {
+func (r *ChatRepo) FindByOrganizationID(orgID uint64, limit, offset int) ([]domain.Chat, error) {
 	var chats []domain.Chat
 	err := r.db.Where("organization_id = ?", orgID).Limit(limit).Offset(offset).Order("created_at DESC").Find(&chats).Error
 	return chats, err
 }
 
 // FindByUserID finds chats by user ID with pagination
-func (r *ChatRepo) FindByUserID(userID uint, limit, offset int) ([]domain.Chat, error) {
+func (r *ChatRepo) FindByUserID(userID uint64, limit, offset int) ([]domain.Chat, error) {
 	var chats []domain.Chat
 	err := r.db.Where("user_id = ?", userID).Limit(limit).Offset(offset).Order("created_at DESC").Find(&chats).Error
 	return chats, err
@@ -56,12 +56,12 @@ func (r *ChatRepo) Update(chat *domain.Chat) error {
 }
 
 // Delete deletes a chat by ID
-func (r *ChatRepo) Delete(id uint) error {
+func (r *ChatRepo) Delete(id uint64) error {
 	return r.db.Delete(&domain.Chat{}, id).Error
 }
 
 // CountByOrgIDAndDateRange counts chats in a date range for an organization
-func (r *ChatRepo) CountByOrgIDAndDateRange(orgID uint, start, end time.Time) (int64, error) {
+func (r *ChatRepo) CountByOrgIDAndDateRange(orgID uint64, start, end time.Time) (int64, error) {
 	var count int64
 	err := r.db.Model(&domain.Chat{}).
 		Where("organization_id = ? AND created_at BETWEEN ? AND ?", orgID, start, end).
@@ -70,7 +70,7 @@ func (r *ChatRepo) CountByOrgIDAndDateRange(orgID uint, start, end time.Time) (i
 }
 
 // GetTagStats gets statistics for tags in an organization
-func (r *ChatRepo) GetTagStats(orgID uint) (map[string]int64, error) {
+func (r *ChatRepo) GetTagStats(orgID uint64) (map[string]int64, error) {
 	// This is a simpler version. For a production implementation,
 	// we'd use jsonb functions in PostgreSQL to extract and count tags.
 	var chats []domain.Chat
